LayoutTests/imported/w3c/web-platform-tests/css/css-values/rlh-invalidation-expected.txt

CONSOLE MESSAGE: Error: assert_not_equals: expect rlh units to update on line-height update got disallowed value 800

FAIL CSS Values and Units Test: rlh invalidation Error: assert_not_equals: expect rlh units to update on line-height update got disallowed value 800
